move dashboard route
