+++ app/routes.js
@@ -6,8 +6,8 @@ import config from './config';
 import { logger } from './logger';
 import { withCatch, getHealthCheckDependencies, extractAccessToken } from './helpers/routes/routerHelper';
 import { getDocumentByFileName } from './helpers/api/dapi/getDocumentByFileName';
-import { getDashboardContext } from './pages/dashboard/controller';
 import { getTaskListPageContext } from './pages/task-list/controller';
+import { dashboardRoutes } from './pages/dashboard/routes';
 import { sectionRoutes } from './pages/sections/routes';
 import { summaryRoutes } from './pages/summary/routes';
 import { completeOrderRoutes } from './pages/complete-order/routes';
@@ -45,16 +45,7 @@ export const routes = (authProvider, sessionManager) => {
     stream.on('close', () => res.end());
   }));
 
-  router.get('/organisation', authProvider.authorise({ claim: 'ordering' }), withCatch(logger, authProvider, async (req, res) => {
-    const accessToken = extractAccessToken({ req, tokenType: 'access' });
-    const context = await getDashboardContext({
-      accessToken,
-      orgId: req.user.primaryOrganisationId,
-      orgName: req.user.primaryOrganisationName,
-    });
-    logger.info('navigating to organisation orders page');
-    res.render('pages/dashboard/template.njk', addContext({ context, user: req.user }));
-  }));
+  router.use('/organisation', dashboardRoutes(authProvider, addContext));
 
   router.get('/organisation/:orderId', authProvider.authorise({ claim: 'ordering' }), withCatch(logger, authProvider, async (req, res) => {
     const accessToken = extractAccessToken({ req, tokenType: 'access' });
